Improved HTML5 parser. Used external dependencies
