+++ src/system/parsers/XML.php
@@ -13,14 +13,10 @@
 
 namespace NovemBit\i18n\system\parsers;
 
-
-use DOMAttr;
 use DOMDocument;
 use DOMElement;
-use DOMNode;
-use DOMText;
 use DOMXPath;
-use NovemBit\i18n\system\parsers\xml\Rule;
+use Masterminds\HTML5;
 
 /**
  * HTML parser with callback function
@@ -47,7 +43,7 @@ class XML
      *
      * @var array
      * */
-    protected $translate_fields = [];
+    protected $query_map = [];
 
     /**
      * Main DomDocument
@@ -62,13 +58,6 @@ class XML
      * @var DOMXPath
      * */
     protected $xpath;
-
-    /**
-     * Main query for Xpath
-     *
-     * @var string
-     * */
-    private $_query;
 
     /**
      * For script tags
@@ -100,14 +89,13 @@ class XML
      * */
     private $_after_translate_callback = null;
 
-    /**
-     * Special encoding fixer string
-     *
-     * @var string
-     * */
-    private $_xml_encoding_fixer = '<?xml encoding="utf-8"?>';
-
     private $_type = self::XML;
+
+    /**
+     *
+     * @var HTML5
+     * */
+    private $_html5;
 
     const XML = 1;
     const HTML = 2;
@@ -115,21 +103,22 @@ class XML
     /**
      * HTML parser constructor.
      *
-     * @param string $xml XML content
-     * @param string $query Xpath Query
-     * @param int $type XML or HTML
+     * @param string        $xml                       XML content
+     * @param array         $query_map                 Xpath Query
+     * @param int           $type                      XML or HTML
      * @param callable|null $before_translate_callback Before init callback
-     * @param callable|null $after_translate_callback After init callback
+     * @param callable|null $after_translate_callback  After init callback
      */
     public function __construct(
         string $xml,
-        string $query,
+        array $query_map,
         int $type = self::XML,
         callable $before_translate_callback = null,
         callable $after_translate_callback = null
     ) {
-        $this->_setQuery($query);
+
         $this->_setType($type);
+        $this->_setQueryMap($query_map);
         $this->_setBeforeTranslateCallback($before_translate_callback);
         $this->_setAfterTranslateCallback($after_translate_callback);
 
@@ -142,9 +131,9 @@ class XML
      *
      * @param string $xml initial HTML content
      *
-     * @return XML
+     * @return self
      */
-    public function load(string $xml): interfaces\XML
+    public function load(string $xml): self
     {
         $this->_setXml($xml);
 
@@ -161,12 +150,28 @@ class XML
     }
 
     /**
+     * @return HTML5
+     */
+    public function _getHtml5(): HTML5
+    {
+        return $this->_html5;
+    }
+
+    /**
+     * @param HTML5 $html5
+     */
+    public function _setHtml5(HTML5 $html5): void
+    {
+        $this->_html5 = $html5;
+    }
+
+    /**
      * Preserve tags
      * Actually using it for preserve script tags that
      * Impossible to parser with DOM parser
      *
-     * @param string $xml Referenced HTML content
-     * @param string $tag Tag that must be preserved
+     * @param string $xml  Referenced HTML content
+     * @param string $tag  Tag that must be preserved
      * @param string $attr Attributes pattern
      *
      * @return void
@@ -201,8 +206,8 @@ class XML
     /**
      * Restoring preserved tags on HTML
      *
-     * @param string $xml Referenced HTML content
-     * @param string $tag Tag That should be restored
+     * @param string $xml  Referenced HTML content
+     * @param string $tag  Tag That should be restored
      * @param string $attr Attributes pattern
      *
      * @return void
@@ -240,89 +245,37 @@ class XML
     }
 
     /**
-     * Fetch current DOM document XPATH
+     * Walk on DOMDocument and
+     * run callback for each `DOMNode`
      *
-     * @param callable $text_callback Callback function for Text Nodes
-     * @param callable $attr_callback Callback function for Attr Nodes
+     * @param callable $callback Callback function with two params
+     *                           1. $node
+     *                           2. $params
      *
      * @return void
      */
-    public function fetch(callable $text_callback, callable $attr_callback): void
+    public function fetch(callable $callback): void
     {
-        $nodes = $this->_getXpath()->query($this->_getQuery());
+        $accept_queries = $this->_getQueryMap()['accept'] ?? [];
 
-        /**
-         * Fetching nodes to get each node in DomDocument
-         *
-         * @var DOMElement $node
-         */
-        foreach ($nodes as $node) {
-            foreach ($this->_getTranslateFields() as $translate_field) {
+        $ignore_queries = $this->_getQueryMap()['ignore'] ?? [];
 
-                /**
-                 * Getting Rule for current set of field
-                 * Then validating node with this rule
-                 *
-                 * @var Rule $rule
-                 */
-                $rule = $translate_field['rule'] ?? null;
+        $ignore_queries = !empty($ignore_queries)
+            ? '[not(' . implode(' or ', $ignore_queries) . ')]' : '';
 
-                if ($rule === null || !$rule->validate($node)) {
-                    continue;
-                }
 
-                $text = $translate_field['text'] ?? null;
+        foreach ($accept_queries as $query => $params) {
 
-                if ($text != null) {
-                    /**
-                     * Fetching child nodes to find Text nodes
-                     *
-                     * @var DOMNode $child_node
-                     */
-                    foreach ($node->childNodes as $child_node) {
-                        if ($child_node->nodeType == XML_TEXT_NODE
-                            || $child_node->nodeType == XML_CDATA_SECTION_NODE
-                        ) {
-                            /**
-                             * Checking if TextNode data length
-                             * without whitespace in not null
-                             * Then running callback function for text nodes
-                             *
-                             * @var DOMText $child_node
-                             */
-                            if (mb_strlen(trim($child_node->data)) == 0) {
-                                continue;
-                            }
-                            call_user_func_array(
-                                $text_callback,
-                                [&$child_node, $text, $rule]
-                            );
-                        }
-                    }
-                }
+            $query .= $ignore_queries;
 
-                $attrs = $translate_field['attrs'] ?? [];
-
-                /**
-                 * Fetching current set attrs and checking
-                 * If node has attribute with this keys
-                 * */
-                foreach ($attrs as $attr => $type) {
-                    if ($node->hasAttribute($attr)) {
-                        $attr_node = $node->getAttributeNode($attr);
-                        /**
-                         * Running callback function for attr nodes
-                         *
-                         * @var DOMAttr $node
-                         */
-                        call_user_func_array(
-                            $attr_callback,
-                            [&$attr_node, $type, $rule]
-                        );
-                    }
-                }
-
-                break;
+            $nodes = $this->_getXpath()->query($query);
+            /**
+             * Fetching nodes to get each node in DomDocument
+             *
+             * @var DOMElement $node
+             */
+            foreach ($nodes as $node) {
+                call_user_func_array($callback, [$node, $params]);
             }
         }
     }
@@ -349,6 +302,7 @@ class XML
         $this->dom = $dom;
     }
 
+
     /**
      * Initialization DomDocument and Xpath
      * Preserving tag that must be restored on save
@@ -359,72 +313,53 @@ class XML
     {
         $xml = $this->_getXml();
 
-        if ($this->_getType() === self::HTML) {
-            $this->_addPreserveField(
-                'script',
-                '(?!\stype="application\/ld\+json")+?'
-            );
-        }
-        foreach ($this->_getPreserveFields() as $field) {
-            $this->preserveField(
-                $xml,
-                $field[0],
-                $field[1]
-            );
-        }
+        //        if ($this->_getType() === self::HTML) {
+        //            $this->_addPreserveField(
+        //                'script',
+        //                '(?!\stype="application\/ld\+json")+?'
+        //            );
+        //        }
+        //        foreach ($this->_getPreserveFields() as $field) {
+        //            $this->preserveField(
+        //                $xml,
+        //                $field[0],
+        //                $field[1]
+        //            );
+        //        }
 
         $this->setDom(new DomDocument());
 
         if ($this->_getType() === self::HTML) {
-            $this->getDom()->preserveWhiteSpace = false;
-            $this->getDom()->formatOutput = true;
+            //            $this->getDom()->preserveWhiteSpace = false;
+            //            $this->getDom()->formatOutput = true;
+
+            $this->_setHtml5(
+                new HTML5(
+                    [
+                        'encode_entities' => false,
+                        'disable_html_ns'=>true,
+                    ]
+                )
+            );
+
+            $this->setDom($this->_getHtml5()->loadHTML($xml));
             /**
              * Set encoding of document UTF-8
              * */
-            @$this->getDom()->loadHTML(
-                $this->_xml_encoding_fixer . $xml,
-                LIBXML_HTML_NOIMPLIED | LIBXML_HTML_NODEFDTD
-            );
-            $this->getDom()->encoding = 'utf-8';
+            //            @$this->getDom()->loadHTML(
+            //                $this->_xml_encoding_fixer . $xml,
+            //                LIBXML_HTML_NOIMPLIED | LIBXML_HTML_NODEFDTD
+            //            );
+            //            $this->getDom()->encoding = 'utf-8';
 
-        } elseif ($this->_getType() == self::XML) {
+        } else {
+            $this->setDom(new DomDocument());
 
             @$this->getDom()->loadXML($xml);
         }
 
         $this->setXpath(new DOMXpath($this->getDom()));
 
-    }
-
-    /**
-     * Getting translate fields set
-     *
-     * @return array[]
-     */
-    private function _getTranslateFields(): array
-    {
-        return $this->translate_fields;
-    }
-
-    /**
-     * Adding translate fields
-     *
-     * @param Rule $rule Rule object
-     * @param string $text Text node type to translate
-     * @param array $attrs List of attributes that must be translated
-     *
-     * @return void
-     */
-    public function addTranslateField(
-        interfaces\Rule $rule,
-        string $text = 'text',
-        $attrs = []
-    ) {
-        $this->translate_fields[] = [
-            'rule' => $rule,
-            'text' => $text,
-            'attrs' => $attrs
-        ];
     }
 
     /**
@@ -452,27 +387,10 @@ class XML
         }
 
         if ($this->_getType() === self::HTML) {
-            $xml = $this->getDom()->saveHTML();
+            $xml = $this->_getHtml5()->saveHTML($this->getDom());
         } else {
             $xml = $this->getDom()->saveXML();
         }
-
-        foreach ($this->_getPreserveFields() as $field) {
-            $this->restorePreservedTag($xml, $field[0], $field[1]);
-        }
-
-        if ($this->_getType() === self::HTML) {
-            /**
-             * Remove <?xml.. syntax string
-             * */
-            $xml = preg_replace(
-                '/' . preg_quote($this->_xml_encoding_fixer) . '/',
-                '',
-                $xml,
-                1
-            );
-        }
-
 
         return $xml;
     }
@@ -487,28 +405,6 @@ class XML
     private function _setXml(string $xml): void
     {
         $this->xml = $xml;
-    }
-
-    /**
-     * Get Xpath query
-     *
-     * @return string
-     */
-    private function _getQuery(): string
-    {
-        return $this->_query;
-    }
-
-    /**
-     * Set Xpath query
-     *
-     * @param string $_query Query String
-     *
-     * @return void
-     */
-    private function _setQuery(string $_query): void
-    {
-        $this->_query = $_query;
     }
 
     /**
@@ -531,29 +427,6 @@ class XML
     protected function setXpath(DOMXpath $xpath): void
     {
         $this->xpath = $xpath;
-    }
-
-    /**
-     * Adding preserved fields
-     *
-     * @param string|null $tag Html tag
-     * @param string|null $attr Html attribute
-     *
-     * @return void
-     */
-    private function _addPreserveField(string $tag, ?string $attr): void
-    {
-        $this->_preserve_fields[] = [$tag, $attr];
-    }
-
-    /**
-     * Get all preserve fields
-     *
-     * @return array[]
-     */
-    private function _getPreserveFields(): array
-    {
-        return $this->_preserve_fields;
     }
 
     /**
@@ -616,5 +489,21 @@ class XML
     private function _setType(int $type): void
     {
         $this->_type = $type;
+    }
+
+    /**
+     * @return array
+     */
+    private function _getQueryMap(): array
+    {
+        return $this->query_map;
+    }
+
+    /**
+     * @param array $query_map
+     */
+    private function _setQueryMap(array $query_map): void
+    {
+        $this->query_map = $query_map;
     }
 }
